+++ changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - New validator for html lang tag (shortcut: lang)
+- New validator for viewport tag (shortcut: vp)
 
 ## [0.2]